requirements.txt
pandas
yfinance
newsapi-python
vaderSentiment
streamlit
plotly